correcion listar
