+++ src/app/modules/solicitud-vehiculo/components/modal/modal.component.html
@@ -1,14 +1,14 @@
-<span *ngIf="leyenda !== titulo">
-    <div class="text-sm-end">
-        <button type="button" class="btn btn-primary btn-rounded" (click)="openModal(content)">{{titulo}}
-        </button>
-    </div>
-</span>
+<!--<span *ngIf="leyenda !== titulo">-->
+<!--    <div class="text-sm-end">-->
+<!--        <button type="button" class="btn btn-primary btn-rounded" (click)="openModal(content)">{{titulo}}-->
+<!--        </button>-->
+<!--    </div>-->
+<!--</span>-->
 
-<span *ngIf="leyenda === titulo && titulo == 'Detalle'">
-    <button type="button" class="btn btn-primary btn-sm btn-rounded boton-cuadrado mx-1" (click)="openModal(content)"><i
-      class="mdi mdi-eye-outline"></i></button>
-</span>
+<!--<span *ngIf="leyenda === titulo && titulo == 'Detalle'">-->
+<!--    <button type="button" class="btn btn-primary btn-sm btn-rounded boton-cuadrado mx-1" (click)="openModal(content)"><i-->
+<!--      class="mdi mdi-eye-outline"></i></button>-->
+<!--</span>-->
 
 
 <!-- Modal -->
