+++ src/app/modules/solicitud-vehiculo/interfaces/data.interface.ts
@@ -1,72 +1,110 @@
 export interface ISolicitudVehiculo {
-  codigoSolicitudVehiculo: string,
-  fechaSolicitud: string,
-  fechaSalida: string,
-  unidadSolicitante: any,
-  vehiculo: Vehiculo,
-  objetivoMision: string,
-  lugarMision: string,
-  direccion: string,
-  horaEntrada: string,
-  horaSalida: string,
-  cantidadPersonas: number,
-  listaPasajeros: any[],
-  usuario: Usuario,
-  jefeDepto: any,
-  fechaEntrada: string,
-  estado: number,
-  motorista: Empleado,
-  listDocumentos: any[],
+  codigoSolicitudVehiculo: string;
+  fechaSolicitud: Date;
+  fechaSalida: Date;
+  unidadSolicitante: string;
+  vehiculo: IVehiculo;
+  objetivoMision: string;
+  lugarMision: string;
+  direccion: string;
+  horaEntrada: Date;
+  horaSalida: Date;
+  cantidadPersonas: number;
+  listaPasajeros: IPasajero[];
+  solicitante:ISolicitante;
+  nombreJefeDepto: string;
+  fechaEntrada: Date;
+  estado: number;
+  estadoString: string;
+  motorista: IMotorista;
+  listDocumentos: IDocumento[];
 }
 
-export interface Vehiculo {
-  codigoVehiculo: number
-  modelo: string
-  marca: string
-  clase: string
-  color: string
-  year: string
-  capacidad: string
-  estado: string
-  tipo: string
-  foto: any
-  n_chasis: string
-  n_motor: string
+export interface IVehiculo {
+  codigoVehiculo: string;
+  placa: string;
+  modelo: string;
+  marca: string;
+  clase: string;
+  color: string;
+  year: number;
+  fecha_tarjeta: Date;
+  capacidad: number;
+  capacidadTanque: number;
+  estado: number;
+  n_chasis: string;
+  n_motor: string;
+  tipo_gas: string;
+  nombrefoto: string;
+  urlfoto: string;
 }
 
-export interface Usuario {
-  codigoUsuario: number
-  email: any
-  password: string
-  empleado: Empleado
+interface ISolicitante {
+  codigoUsuario: string;
+  email: string;
+  password: string;
+  empleado: IEmpleado;
 }
 
-export interface Empleado {
-  dui: string
-  nombre: string
-  apellido: string
-  telefono: string
-  licencia: any
-  tipo_licencia: any
-  fecha_licencia: any
-  estado: any
-  jefe: boolean
-  correo: string
-  nombrefoto: any
-  urlfoto: any
-  cargo: Cargo
-  departamento: Departamento
+interface IEmpleado {
+  codigoEmpleado: string;
+  dui: string;
+  nombre: string;
+  apellido: string;
+  telefono: string;
+  licencia: string;
+  tipo_licencia: string;
+  fecha_licencia: Date;
+  estado: number;
+  jefe: boolean;
+  correo: string;
+  nombrefoto: string;
+  urlfoto: string;
+  cargo: ICargo;
+  departamento: IDepartamento;
 }
 
-export interface Cargo {
-  codigoCargo: number
-  nombreCargo: string
-  descripcion: string
-  estado: number
+interface ICargo {
+  codigoCargo: string;
+  nombreCargo: string;
+  descripcion: string;
+  estado: number;
 }
 
-export interface Departamento {
-  codigoDepto: number
-  nombre: string
-  estado: number
+interface IDepartamento {
+  codigoDepto: string;
+  nombre: string;
+  estado: number;
+}
+
+interface IPasajero {
+  codigoUsuario: string;
+  nombre: string;
+  apellido: string;
+}
+
+interface IMotorista {
+  codigoEmpleado: string;
+  dui: string;
+  nombre: string;
+  apellido: string;
+  telefono: string;
+  licencia: string;
+  tipo_licencia: string;
+  fecha_licencia: Date;
+  estado: number;
+  jefe: boolean;
+  correo: string;
+  nombrefoto: string;
+  urlfoto: string;
+  cargo: ICargo;
+  departamento: IDepartamento;
+}
+
+interface IDocumento {
+  codigoDocumento: string;
+  nombre: string;
+  descripcion: string;
+  archivo: string;
+  estado: number;
 }
